Flatten database structure now that we plan to handle migrations independently.
